+++ github/messages.go
@@ -45,8 +45,8 @@ var (
 		"deployment_status":                     "DeploymentStatusEvent",
 		"fork":                                  "ForkEvent",
 		"gollum":                                "GollumEvent",
-		"integration_installation":              "IntegrationEvent",
-		"integration_installation_repositories": "IntegrationEvent",
+		"integration_installation":              "IntegrationInstallationEvent",
+		"integration_installation_repositories": "IntegrationInstallationRepositoriesEvent",
 		"issue_comment":                         "IssueCommentEvent",
 		"issues":                                "IssuesEvent",
 		"member":                                "MemberEvent",
